requirements.txt
numpy
matplotlib
jupyter
sympy
scipy
pandas>=0.22
pyyaml>=3.1
filterpy>=1.2.4
seaborn>=0.8.1